Allow using authorization header for /graphql endpoint
Parabol 9.0.0 does not support the x-application-authorization header
anymore. In order to be able to use the authorization header for the
endpoint in the plugin, I had to remove the check that the request is
coming from a logged in Mattermost user. This is not a problem because
we forward the request to a public Parabol endpoint which checks the
authorization anyways based on the authorization header.